rsvg-convert: use render_document() vs. render_element() depending on --export-id
Strictly speaking, this changes the behavior from the C version, as
render_element() renders the untransformed element instead of the
layer as it is (the C code used rsvg_handle_render_cairo_sub(), which
calls render_layer() underneath).

However, I posit that the behavior from render_element() is actually
more useful for the "extract an object" case that people want to use.
The behavior from the C code was broken enough to not be useful.