Fix the export_id_short_option geometry
This one is debatable.  The document has units in inches, and a
viewBox.  In that case, what does it mean to render elements at their
natural size?  Is it at the size of the untransformed element (current
behavior, and what the test expects), or is it the size at which the
element would be rendered if rsvg-convert were run with no argumnents
for the whole document?