Drop temporary values earlier in SpecifiedValues::to_computed_values()
The `prop_val` in each of the `compute!` expansions would previously
only be dropped at the end of the function although it's only used
inside the macro. To prevent that, add a new block inside the macro to
have it dropped at the end of each macro expansion.

Maybe fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/686