Don't cache resolved gradients anymore
We will resolve the stroke/fill opacity directly into the gradient.
Since this is different for each invocation of the gradient (for the
stroke *or* fill *or* the current color at each element's cascade), we
cannot keep a single cached copy of the resolved gradient.

Maybe this will become possible again if we ever cache
UserSpacePaintSource in a render tree or something.