Store a pattern's intended opacity in the ResolvedPattern / UserSpacePattern
This is similar to baking gradients together with their final
opacity.  A pattern's opacity is not applied until the end, when
painting its pattern surface with the opacity, unlike a gradient's,
which can be built into each stop's opacity.

If we ever switch to generating gradient surfaces (analogous to
pattern surfaces), we can handle both in the same way.