ci: Port to fd.o ci-templates
Till now we used a different repo to build and manage docker images. [1]
The images where set to rebuild on their own to reduce maintainance
and cross repo synchronization.

Latetly we have started to transition to CI-Templates [2] which can manage
and build images from a single template and inline the repo.

In order to edit/rebuild the images bump the base tag and modify the
FDO_DISTRIBUTION_PACKAGES and FDO_DISTRIBUTION_EXEC variables as needed
in .gitlab-ci/container_builds.yml. Then you can use the generated
images in the jobs at .gitlab-ci.yml.

Refer to the ci-templates documentation for more. [3]

[1] https://gitlab.gnome.org/GNOME/librsvg-oci-images
[2] https://gitlab.freedesktop.org/freedesktop/ci-templates
[3] https://freedesktop.pages.freedesktop.org/ci-templates/templates.html