Pass the color_interpolation_filters directly to FilterContext.get_input()
There was a convoluted scheme to get the value of
color_interpolation_filters in the main filters::render(), and hack it
into the FilterContext with with_linear_rgb().

It is easier if filter primitives request their inputs in the color
space that they need.

I've added comments where the spec explicitly says that a primitive
should *not* use the value of the color_interpolation_filters
property.

I've also added a comment to morphology.rs, since the spec is not
clear about it, but a) using the value from the ComputedValues makes a
test fail; b) the original code has feMorphology's implementation of
is_affected_by_color_interpolation_filters() set to false anyway.

I've regenerated the reference for svg1.1/filters-displace-02-f.svg -
it fails when the code works as per the spec (it didn't before, as it
converted both `in` and `in2` to the color_interpolation_filters color
space when it should only do that for `in2`, while `in` should remain
in its original color space.  I can't make much sense of the
difference between the new result and the original PNG from the SVG1.1
test suite.