Remove pangoft2 depdency from C test code
The tests written in C used to set up the fontmap in order to get
reproducible results for reference tests. Since these tests are now
part of the Rust test suite, and the remaining tests written in C
do not exercise text rendering in reference tests, this code isn't
needed any longer.

This remove the dependency on pangoft2 from the C code, thus paving
the way for !472.