Move the test for #347 to Rust
This is a test for a particular layer's position; #347 was about text
elements not getting their bounds computed correctly.

Note that the original test in C had this:

    if (fixture->has_position) {
        g_assert_cmpint (fixture->x, ==, position.x);
        g_assert_cmpint (fixture->y, ==, position.y);
    }

Even though fixture->x is a double, it is being cast to an int to
compare it to position.x (RsvgPositionData uses ints).

The test effectively does "is the position.y between 48 and 49", so
that's what we use in the Rust test.