(#703): Ignore elements in an error state inside the <switch> element
It turns out that the implementation for <switch> needs to be aware of
elements that have been set in an error state while parsing their
attributes.  For example, getting an error while parsing
`systemLanguage` for an element inside a <switch> should cause *that*
element to be ignored while picking which of `<switch>`'s children to
render.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/703