Fix which computed values are used to resolve lengths in the filter element and primitives
Per my understanding of the spec
https://www.w3.org/TR/filter-effects/#FilterPrimitiveSubRegion

The x/y/width/height for the <filter> element, and for all <feFoo>
filter primitive elements, are in the coordinate system determined by
the values of filterUnits and primitiveUnits, respectively.  That is,
they should use the current viewport or the current element's bounding
box depending on those values.

However, font-relative units like "em"/"ex" resolve with respect to
the font-size of the element where they are used (i.e. the <filter> or
<feFoo>), *not* the element that referenced the <filter>.  I think
Firefox handles this incorrectly.

In fact, the test filter-effects-region.svg had a comment along those
lines - only it considered Firefox's behavior to be correct, and
Chromium to be in error.  I think it is the other way around, and have
done two things:

* Tweaked the values in that test file to be easier to eyeball.

* Regenerated the reference file per my understanding of what should
  happen in the "cascading" and "cascading-primitive" cases (upper
  right and center right, respectively).

This also makes it obvious that the computed_from_node_being_filtered
is not needed at all in the filters code.