Do not constrain units for x/y/width/height in filters based on objectBoundingBox
I can't find anything in the spec that indicates that using
objectBoundingBox for filterUnits or primitiveUnits means that
x/y/width/height can only be numbers or percentages.  I interpret the
spec to mean that for objectBoundingBox, well, those lengths *are*
normalized with respect to the bbox of the object being filtered.

This changes the test for filter-effects-region.svg once again.  The
reference still disagrees with WebKit / Firefox; I'll file a bug about that.