(#698): Put limit on feMorphology radii to prevent thread hang
Limits feMorphology radii to 10.0, a somewhat arbitrary number that I attempted to balance quickly (on a Ryzen system) opening eog and having a wide range for the effect.
The examples for radius size I could find only had it go as high as 3, so I figured 10 would be safe.
If there's something I've done wrong, or how I implemented it is odd for rust's conventions please let me know!

https://gitlab.gnome.org/GNOME/librsvg/-/issues/698