Translate the blur() filter function into filter primitives
This basically involves constructing a FilterSpec by hand, mostly out
of default values, and a PrimitiveParams::GaussianBlur.

The test for this renders two SVG documents.  On one, an element is
modified using a blur() filter function.  On another, the reference,
the same element is modified with a <filter> element with a
<feGaussianBlur> - the same expansion for the filter function as the
spec suggests.