DrawingCtx.nested() - replaces push_cairo_context / pop_cairo_context
The cr_stack is just used to keep a stack of surfaces for compositing
and their transforms (for compositing in get_snapshot(), used by the
filters code).

In reality, push_cairo_context / pop_cairo_context were used as a way
to paint to another cr, that comes from another surface.

Now, we can create a "nested" DrawingCtx with respect to the other cr.
Maybe this will let us unify DrawingCtx / surface / cr creation.