length::NormalizeParams - new struct to hold the values for length normalization
Everything that needs to normalize lengths needs to carry around a
ComputedValues for the font size, and a ViewParams for the current vbox
and DPI.  Finally, these get passed to length.normalize().

I want to see if using a single struct NormalizeParams lets us avoid
passing the DrawingCtx everywhere in callers - it's needed just to get a
ViewParams in some uncomfortable places.