Remove DrawingCtx::with_saved_transform(); pass the transform to with_discrete_layer()
Element::draw() used to set the draw_ctx's transform itself by calling
DrawingCtx::with_saved_transform().  So, the code flow was this:

  Element.draw() {
      with_saved_transform(
          element.transform,
          {
              element_impl.draw() {
                  draw_ctx.with_discrete_layer({ ... })
              }
          }
      )
  }

Now, we pass the transform to use directly to with_discrete_layer().
The code flow becomes this:

  Element.draw() {
      element_impl.draw() {
          draw_ctx.with_discrete_layer({ element.transform, ... })
      }
  }

This requires two notable changes:

1. draw_from_use_node(): must now apply the <use> element's transform
in the same fashion as with_discrete_layer().

2. "impl Draw for Text" is what starts the recursive process of
creating text chunks and spans.  It used to be that generating the
chunks/spans was done outside the call to
DrawingCtx::with_discrete_layer().

However, now that the draw_ctx.cr is *not* transformed until the call
to DrawingCtx::with_discrete_layer(), I had to move the code that
creates the chunks/spans to be inside the call to
with_discrete_layer().  This is because creating a Pango layout
requires the cr to already have its final transform.