New StackingContext struct
Let's create a layout tree from the DOM tree!

This commit introduces StackingContext, or "all the shit that
DrawingCtx::with_discrete_layer() pulls out of the node being rendered
and the ComputedValues".  That information is what the SVG spec says
would be necessary to determine whether an element establishes a
stacking context, to be rendered as an isolated group or not -
basically librsvg's needs_temporary_surface flag.

For now, StackingContext just holds the transform.  I'll be moving the
other parameters into it in separate commits.