(#704): Circle/ellipse made out of a single Arc command does not work
https://www.w3.org/TR/SVG2/paths.html#ArcOutOfRangeParameters says "If
the endpoint (x, y) of the segment is identical to the current
point (e.g., the endpoint of the previous segment), then this is
equivalent to omitting the elliptical arc segment entirely.", but this
is inconsistent with shapes-intro-02-f.svg from the SVG 1.1 test
suite.  Also, tools in the wild are generating filled circles by doing
just "Moveto Arc", with the end points of the arc being the same as
the Moveto's.

However, nothing in the endpoint-to-center parameterization algorithm
in https://www.w3.org/TR/SVG2/implnote.html#ArcConversionEndpointToCenter,
or in the subsequent section for out-of-range radii
https://www.w3.org/TR/SVG2/implnote.html#ArcCorrectionOutOfRangeRadii,
mention ignoring the Arc segment if its start/end endpoints are
equal.  The algorithm still works.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/704