Store LanguageTags inside SystemLanguage and really evaluate them at rendering time
Instead of storing the result of evaluating the systemLanguage
attribute at loading time, SystemLanguage is now a newtype around a
set of LanguageTags.  It is matched against the environment's language
tags at rendering time.

And at rendering time, it is now DrawingCtx which generates the list
of the user's language tags from the environment.

I remove the Lazy<Locale> - the tests started failing with the
shuffling above.  The result of g_get_language_names() will only be
converted once per render, anyway, because now it is done in DrawingCtx.