(#746) - Possible cairo_save() without cairo_restore() in render_layer()
The code has a "?" in the call to get_node_or_root, so it could return
before doing the cairo_restore().

This is not a problem for rsvg-convert, since it always uses
render_element when an element id is specified.

It is critical for the C API, where render_cairo_sub/render_layer call
into Handle.render_layer.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/746