(#677) - Do not clip the rightmost/bottomost pixels of an image with partial pixel coverage
In rsvg-convert, to compute the pixel size of PNG output, we now
ceil(size) instead of round(size).  This prevents clipping of the last
row or column.

This is also in place now for the legacy functions that output to pixbufs.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/677

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/543>