Use the smcp (small caps) OpenType feature for fonts that support it
This is part of #668.  It turns out that Pango does not implement
PANGO_VARIANT_SMALL_CAPS.  Instead, one must turn on the "smcp"
OpenType feature, but not all fonts support that.  In the future we
may need to synthesize small caps on the librsvg side if Pango doesn't
do it itself.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/545>