New test for drawing to a non-image surface in the Rust tests
This is slightly different from the one in tests/api.c - that one
tests that the cr's transform is preserved.

We'll leave that one as it is, and use the Rust one to test for the
tricky code paths in drawing_ctx.rs.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/553>