(#758): Panic when rendering with masks or opacity to a non-image surface
Change the non-image-surface test to exercise a untested code paths.

This exposes this bug:

  thread 'bugs::can_draw_to_non_image_surface' panicked at 'called
  `Result::unwrap()` on an `Err` value: Surface(0x7f1728022bd0)',
  src/drawing_ctx.rs:741:88

in the unwrap() here:

    let surface_to_filter = SharedImageSurface::copy_from_surface(
        &cairo::ImageSurface::try_from(temporary_draw_ctx.cr.target()).unwrap(),
    )?;

This happens in the case where there are no filters, but there is
opacity, and the code incorrectly assumes that it has an image surface
to work with.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/758

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/553>