(#759) - Replace SavedCr again - use a with_saved_cr() function
Sigh, it was so nice to cr.restore() on Drop of the SavedCr, but since
restore() can now return an error, this is no longer practicable.  So,
we go back to the scheme of using a helper function - just so we can
use the ? operator easily in the callbacks.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/759

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/559>