rsvg-convert.rs: Fix building on Windows
Make sure we import the things we really need, and like what is
happening on *nix, the Win32 HANDLES that we get for stdin/stdout
need to be marked as unsafe before we can cast the resulting items.

Also make sure we deal with OutPutStream with the corresponding
Win32OutputStream.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/565>