NMake Makefiles: Update for cargo-c usage
This updates the Visual Studio build files to build librsvg after it has
been updated to use cargo-c to build the C libraries, and, as a
side-effect, static builds are now supported as well at the same time.

As a convenience, we still provide rsvg-2.0-vsXX.dll resulting from the
builds for people that aren't able to re-link their applications against
rsvg-2.dll (the DLL file name that results from the cargo-c updates),
but a warning message is shown upon 'install' that people are advised to
re-compile their applications if possible, since rsvg-2.0.lib now links
to rsvg-2.dll.