NMake Makefiles: Streamline build process
Use separate build directories for librsvg and rsvg_convert, so that
building one of it does no require rebuilding the other.  Also make the
cross-build batch generation done on the fly, so that they will be
removed once the Rust cross build work is completed (or failed--they can
be generated quite quickly if needed).