(#771): Fix incorrect default for the flood-color property
Per the spec, the default is opaque black.  However, librsvg's default
value was all-zeros, or transparent black.

43499a14af1ad06620d500462ac0c8ad6fd8bec5 refactored the feFlood
primitive to pass an already resolved color to
SharedImageSurface::flood().  Prior to that, the code in flood() was
incorrectly using just the RGB values from the flood-color property,
but an alpha value equal to the flood-opacity - instead of first
resolving the whole RGBA flood-color with the flood-opacity.

Now that the flood-color is resolved properly, it turns out that the
default was wrong.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/771

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/573>