impl Parse for Transform: don't ensure that it is invertible
This is a general change in strategy.  We parse e.g.
matrix(0, 0, 0, 0, 0, 0) as a Transform, but leave it up to the caller
to ensure that the transform is usable.

This happens in drawing_ctx, where it uses the transform from the
StackingCtx.  Per the CSS Transforms spec, objects should just not be
rendered if the transform is not invertible, but that condition should
not set the element to be in error.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/576>