text_anchor_advance: take a horizontal flag, not the whole WritingMode
This is the "horizontal" flag from laying out the whole <text>
element, per the text layout algorithm in the SVG2 spec.  I'm starting
to question it, though - what if there is embedding of vertical within
horizontal text?  Shouldn't it be per-chunk or per-span as usual?

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/593>