Implement `:link` pseudo-class
This also adds parsing for the `:visited` pseudo class,
but since librsvg maintains no history, it's not much of an implementation.

This change is included to make the test case in #738 pass,
which doesn't just try to match an `a` selector, but a whole
`a:link` selector.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/601>