StackingContext: store the link_target here for <a> elements
The <a> Link element used to draw itself by wrapping its drawing in a
call to DrawingCtx::with_link_tag().  We remove that function, pass
the link target as part of StackingCtx, and deal with it in
DrawingCtx::with_discrete_layer().

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/613>