Use the objectBoundingBox of the whole text element for each span's individual paint source
Per the spec, if a <tspan> has paint that depends on objectBoundingBox, that box
corresponds to the whole <text> element, not to the individual <tspan>.

Now we do this:

1. PositionedSpan::layout() -> LayoutSpan, which has resolved paint
sources, but not in user-space coordinates yet.

2. Combine all the bboxes from the LayoutSpans to find the whole text_bbox.

3. Convert the paint sources to user space with the text_bbox; produce
TextSpans; gather them in a layout::Text.

4. Send off the layout::Text to DrawingCtx::draw_text().

In the future, layout::Text will contain per-glyph positions so we can
implement more features of the SVG text spec.

This commit also adds a test using the Ahem font, to make a reftest
with plain squares for each glyph.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/613>