(#804) - tspan that changes the text direction is placed incorrectly
The code was using the direction from the span, not the direction of
its parent chunk, and it is the chunk's direction that lays out its
child spans.

I don't think this will handle nested <tspan> that change directions,
but that's for later.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/804

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/615>