Collect the advance of a chunk's spans until PositionedChunk::from_measured
I.e. don't store the collected advance of a chunk's spans in
MeasuredChunk.  That advance does not make sense; it's just the way
the code computes things right now.  Let's move it to the place that
can do anything meaningful with that information, and refactor from there.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/616>