Extract the dx/dy from the first span of a chunk and make it the chunk's own dx/dy
Librsvg does not yet support arrays of dx/dy values in those
attributes.  So, the single value they support acts as if it overwrote
the one from the parent element.  That is, in

  <text dx="-1">
    <tspan dx="2">
      ...
    </tspan>
  </tspan>

The only dx ever used will be dx=2, since it would override the -1.
Here, dx=2 becomes part of the first and only chunk.

At the end, we add the chunk's dx/dy to each positioned span,
independend of the chunk's internal layout.

This finally makes
https://upload.wikimedia.org/wikipedia/commons/f/f7/S-Bahn_RheinNeckar2020.svg
render correctly.  Yay!!!

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/616>