(#249): Basic implementation of the unicode-bidi property
This commit causes each span's text to be wrapped with the directional
control characters specified in
https://www.w3.org/TR/css-writing-modes-3/#unicode-bidi

For a more complete implementation of unicode-bidi, we'll have to
change the code to actually bidi/shape the whole character content of
a <text> element and its children, not just each individual span.
That will come soon.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/621>