DrawingCtx.initial_transform_with_offset - remove this function; use the plain initial transform
Now that draw_tree() creates an initial transform and viewport so that
the latter has (0, 0) at the origin, we don't need to offset the
initial transform passed to CompositingAffines::new().  We can use the
plain initial transform as computed by draw_tree() at the beginning of
the drawing process.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/629>