Update to selectors 0.23.0 and cssparser 0.28.0 together
This is the latest version of selectors.
This is not the latest version of cssparser; the latest is 0.20.

However, selectors 0.23.0 depends on cssparser's traits, so they need to be
compatible versions.  This is why we only update to cssparser 0.28.0.

This commit is mostly about adjusting to the API changes in selectors 0.23.0:

* Several of the associated types in the selectors::SelectorImpl trait
now must implement cssparser::ToCss instead of
fmt::Display (apparently to save allocations).

This requires us to use newtypes around the lower-level types we were
using for those associated types:

  struct AttributeValue(String);
  struct Identifier(markup5ever::LocalName);
  struct LocalName(markup5ever::LocalName);
  struct NamespacePrefix(markup5ever::Prefix);

Each of those newtypes implements From<&str> and ToCss.  Specific
types have more specific requirements, like `impl AsRef<str> for AttributeValue`.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/641>