Visual Studio: Support building from GIT checkout
Add rules in win32/generate-msvc.mak to create win32/config-msvc.mak,
win32/config.h.win32 and include/librsvg/rsvg-version.h so that we can
support building from a GIT checkout using Visual Studio 2013+ without
manually editing the *.in files for the build.  Add also a rule to
remove the generated files and clean up the vsXX build directories.
These rules are meant for GIT checkouts only and are not meant for
release tarballs, at least for now.

The newly-added win32/prebuild.py is not meant to be distributed, so it
is not added to 'make dist'.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/653>