Hack around grcov bug to make the toplevel index.html actually show all source directories
Found a bug in grcov!
https://github.com/mozilla/grcov/issues/753

Normally, grcov creates this:

  html/             substituted with the --output-path option's value
    index.html      index of all directories with their coverage summaries
    src/
      index.html    summary of all source files in src/
      foo.rs.html   report for src/foo.rs
      bar.rs.html   report for src/bar.rs

However, the presence of build.rs in the toplevel source dir causes
grcov to overwrite its toplevel html/index.html (the index of all
directories) with the report for top_srcdir/index.html (the summary
for the toplevel source directory, that only lists build.rs).

So, we ignore build.rs for now in the HTML coverage report.  I would
also like to get coverage info for build.rs, but it's not terribly
important right now.  The only uncovered source lines in it are fatal
error handlers, the ones that would cause the build to fail anyway.