Remove obsolete tests for 777834-empty-text-children.svg
1. This was one of the tests that we started ignoring recently, when
Pango reverted its fix for for fonts that are not scaled the same in
the x/y dimensions.

2. That test was for the old
https://bugzilla.gnome.org/show_bug.cgi?id=777834, where the Rust code
for patterns would panic when the C code for text generated an empty
bounding box.

Since the code for patterns in (2) is fixed since a long time ago, we
no longer need this test.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/660>