NMake Makefiles: Fix introspection for release tarballs
We were checking unnecessarily whether we are building from a GIT checkout for
the various items that we use in the NMake Makefiles, so add a check when
invoking the NMake Makefiles before we check for the GIT-only items.

Also, make the process of generating the file listings for the introspection
more silenced.

This will fix the introspection builds from release tarballs.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/662>