Use the AspectRatio machinery for ResizeStrategy::Fit when keeping the aspect ratio
Along with this, fix a broken test in the rsvg-convert command-line
tests.  Rendering an SVG with proportions 1:4 with

  rsvg-convert --width=500 --height=1000 --keep-aspect-ratio one-to-four.svg

should produce a 250x1000 pixel image!

This means that the code to scale proportionally in rsvg-convert was
generally broken :(

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/669>

Backport of 462cdb92626e79034302a17a8828ff3c509e85dd