NMake Makefiles: Support generating documentation
Since the switch to gi-docgen, it is much easier to generate documentation for
Visual Studio builds, so add support for that in the NMake Makefiles.

Generally, it depends on the Python module gi-docgen to be installed (quite
easily via PIP), and requires that one passes in INTROSPECTION=1 along with
generate-docs in the NMake command line.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/694>