(#823): Regression in geometry computation when the SVG viewBox does not match the default viewport
Curiously, this manifests in the same way as #521, and is fixed by
removing the code that I added to fix that one.

I'm thinking that all the cleanups around viewport and viewBox
handling caused the old code to not be needed anymore.

(Found by adding the same debug output to librsvg-2.48 and the main
branch, and comparing executions.)

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/823

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/695>