Disable the cobertura.xml report for now
Disable the cobertura report for now; it is only used for showing coverage
in the diff view of merge requests.

After switching to gcov-based instrumentation (-Zprofile above), this
coverage.xml is almost 500 MB and causes gitlab's redis to OOM when
transfering that artifact from the runner to the main gitlab instance.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/725>