Remove ClipMode::ClipToVbox, it is not used anywhere
Commit 62f31feec7 mentions
https://www.w3.org/TR/SVG11/masking.html#AutoClipAtViewportNotViewBox
- but now I think I misunderstood it.  If one wishes to clip to the
viewBox, that version of the spec mentions setting the `clip` property
to the same bounds as the viewBox.

However, the `clip` property is deprecated (Appendix A: The deprecated
clip property)- https://drafts.fxtf.org/css-masking/#clip-property and
librsvg never implemented it anyway.  Looks like we'll have to
implement that appendix, or use clip-path instead.

Also, the clip mode was passed around as Option<ClipMode> - reduce it
to an enum with two cases (more legible than a boolean), and remove
the Option.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/737>