Xml2Parser: store a plain reference to XmlState, not an Rc<XmlState>
Now that I know how to use structs with reference fields and lifetimes...

The lifetime of the Xml2Parser is basically the duration of
XmlState::parse_from_stream(), and it just calls the XmlState to add
elements and such.  Let's encode that into the Xml2Parser struct.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/752>