XmlState: don't take() the document_builder in the end; just consume everything
DocumentBuilder.build() consumes its self argument, which makes sense:
it turns the builder into a document, and that's that.

However, XmlStateInner had an Option<DocumentBuilder> field just so
that the option could be take()n in the end and then the builder
could be consumed by .build().

Using an Option<DocumentBuilder> is inconvenient because the rest of
the code must then do inner.document_builder.as_mut().unwrap()
everywhere.  We could have used a helper function to do that, but we
can remove that construct altogether.

One thing to note is that since we are consuming the final XmlState
and just pick out its inner field in build_document(), we can no
longer impl Drop for XmlState to free the hash table of libxml2
entities.  So, we inline the code to free the hash in
build_document() instead.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/752>