extract_filter_from_filter_node(): Don't take the DrawingCtx; take the possible ViewParams instead
The only reason that function takes a DrawingCtx argument is to ask
the draw_ctx for the view_params depending on the filter units and the
primitive units (... and also, to get the Session).

Instead, create a new ViewParamsGen struct (for "generator" or
something) that holds the two possible ViewParams values for those
units - and pass that struct to the function.

I think this may see some use elsewhere.  The only other place that
calls draw_ctx.get_view_params_for_units() is generate_cairo_mask().

This goes in the direction of converting everything to user space as
soon as possible, instead of as late as possible.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/753>