FilterContext::get_raw_input() - If a filter input is not found, fall back correctly per the spec
E.g. <feSomePrimitive in="nonexistent">

We used to return FilterError::InvalidInput, which would invalidate
the entire filter chain.  Per the spec, when a named "in" or "in2" is
not found, we need to act as if that input were unspecified - i.e. the
last result in the filter chain, or SourceGraphic for the first
primitive in the chain.

This fixes the mouse's hat in https://gitlab.gnome.org/GNOME/librsvg/-/issues/912