Add a test for non-positive stdDeviation in feGaussianBlur
This is https://gitlab.gnome.org/GNOME/librsvg/-/issues/913 and it
already works fine, but more or less by accident - the feGaussianBlur
filter primitive gets marked as "in error" due to the negative
stdDeviation, and the filters code skips it.

This test is for future-proofing the refactoring that follows.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/763>