SystemLanguage: turn into an enum; store Valid and Invalid variants
Previously, parsing an invalid BCP47 language tag would be reported as
an Err(), and the corresponding element inside a <switch> would be set
"in error".  Then, the <switch> implementation would detect that the
child element is in error and ignore it.

We are switching to storing the error condition in the SystemLanguage itself.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/763>