(#928): Fix panic when a feTile filter gets an empty source region
When the input_bounds from the filter source is empty, we tried to
create a zero-sized image for tiling.  This is not allowed.  There is
nothing to tile in that case, so just return the original image to filter.

Fixes https://gitlab.gnome.org/GNOME/librsvg/-/issues/928

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/787>