Copy the sources of rsvg-bench to here.
For Shinigami's Outreachy internship, we'll need a way to have a
benchmark that uses an exact librsvg commit.  So, instead of messing
around with cargo options to use an exact commit for the crate, let's
build the benchmark into the librsvg source tree.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/789>