gradient::Common - don't impl SetAttributes, use a simple method
In 6608c133d501bca6ffe2d1288539c91882fe5bdb, ElementTrait was split
into two traits, SetAttributes and Draw.

The purpose of SetAttributes back then was just to be able to
implement it for gradient::Common, which is a simple struct, not an
SVG element.

Since it is *not* an element, let's not use that trait for it at all.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/801>