filters/component_transfer.rs: Remove the channel field in FeFunc*
This is a vestige of the C code.  All the feFuncX elements were
represented with a single struct that had a "channel" field, to
identify which is which.

  <feComponentTransfer ...>
    <feFuncR .../>
    <feFuncG .../>
  </feComponentTransfer>

When scanning the children of feComponentTransfer, the individual
func->channel were tested to see which of the children is the Red
function, which the Green function, etc.

But now we identify the individual functions by their element type, so
we can just do `matches!(element, Element::FeFuncR)` or similar.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/801>