New function to get the transform that will be used for the stacking context
This fixes @notriddle's MR when rebased on top of the changes to have
a ValidTransform instead of just Transform in DrawingCtx.

The text code needs to know the actual transform that will be used on
the target surface, which could be the temporary surface for
compositing an isolated stacking context, or the "normal" surface
without isolation.

This is the transform that is passed to the draw_fn callback in with_discrete_layer().

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/816>