NormalizeValues: Hacky (?) way to extract what is needed from ComputedValues for normalization
I want to remove the use of ComputedValues from
DrawingCtx::with_discrete_layer(), but it is needed for
NormalizeParams::new().  So, provide a temporary type NormalizeValues
that just extracts the information needed from ComputedValues - in
this case, just the font size.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/816>