Remove the test filter-morphology-from-reference-page.svg
This was added during the port of the filters code to Rust, but it is
redundant with tests/fixtures/reftests/svg1.1/filters-morph-01-f.svg

Also this file produces frequent failures in the test suite whenever
the font rendering stack changes even just a bit.  The point of the
test is to test the feMorphology, not text rendering - so remove it in
favor of the other test from the SVG1.1 test suite, which does erosion
and dilation filters on paths, not on text.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/817>