Combine ViewParams and Viewport into a single struct Viewport
ViewParams was the original set of values that we used since 2018 for
length normalization.

Viewport was my attempt to start unifying the "viewport" concept,
around 2020 and #298.

My current plan is to remove the current viewport from the mutable
part of DrawingCtx, and instead pass the current viewport as function
arguments to the draw() functions.

This commit basically moves the fields and methods from ViewParams
into Viewport, and makes everything use Viewport instead of
ViewParams.  Things are a bit funky because of the viewport_stack
field, which is shared via an Option<Weak<>> to the viewport stack in
the DrawingCtx - but that will disappear when the current viewport is
passed as arguments instead of mutable state.

Part-of: <https://gitlab.gnome.org/GNOME/librsvg/-/merge_requests/821>